Refactor: Product, ProductsModel, sync. with Firebase.
Instead of of a spinner and keeping the user waiting, keep a local
database (will be SQLite) and sync in the background automatically.

Currently sync is to be triggered manually for add, update and
automatically for delete. Will need to decide what is the best strategy
for background sync: periodically or when the products are fetched ?!